requirements.txt
selenium==4.1.0
pytest==7.1.2
allure-pytest==2.9.45
